Tried to implement boundry stencils but realized it was not trivial how to do it.
--HG--
branch : feature/stencils
